add a quick integration test with ActiveRecord::Base
a few things to add to the adapter before it can even be used by ActiveRecord:

you need to define a method called "#{adapter_name}_connection" and it
should return the connection adapter you're making. I had to trace through
the whole active_record repository to figure this out.

It was this block (active_record/connection_handling.rb:47):

    def establish_connection(spec = nil)
      spec     ||= DEFAULT_ENV.call.to_sym
      resolver =   ConnectionAdapters::ConnectionSpecification::Resolver.new configurations
      spec     =   resolver.spec(spec)

      unless respond_to?(spec.adapter_method)
        raise AdapterNotFound, "database configuration specifies nonexistent #{spec.config[:adapter]} adapter"
      end

      remove_connection
      connection_handler.establish_connection self, spec
    end

The result of `spec.adapter_method` is dynamically set to what the user
passed in as `{adapter: whatever}` and in this case, will result in calling
`self.whatever_connection`

Since the module ConnectionHandling is mixed-in ActiveRecord::Base, I
had to reopen that module and define the method `gsheets_connection` and
return the adapter that I was just trying to make.

Quite convoluted and not as straight-forward :(