add convenience task of doing offline authentication
i tried using the official google drive library, but it was horribly
confusing. i found it much easier to dive down to the basic authenticator
that the library was using and get the authentication codes from there.

i will probably still  use the google drive library for most of the drive
work since they have already abstracted away much of the stuff i need to
do, but it depends. if it's as convoluted as the authentication methods
(i jope not) then i might have to just roll my own.

another argument for rolling my own is that the api client includes almost
all the google apis, and im just really interested in the drive api.