finally figure out how to get the access token from signet
also made available a few methods to retrieve the refresh token out from
the signet client. you will want to retrieve this in order to save it
later on for a future call.

i also included a sample database.yml entry that includes some of the
options that i think will be used for confuguration.