derive from a fresh abstract adapter interface and reverse engineer our way up
the issue with deriving straight from AbstractAdapter is that there is so much magic going on.
AbstractAdapter is supposed to be a simple interface to help implement any adapter; however it
does not behave as just an interface. it actually mixes in a number of implementation details that
aren't needed for databases that don't conform to the type of databases ActiveRecord natively
supports (i.e. SQL databases).

this makes it very problematic to implement an adapter that doesn't use SQL (in this case, Google
Sheets). at least in my mind, an ActiveRecord adapter allows you to use varying degrees of the
ActiveRecord DSL (e.g. User.first, User.count, User.posts.where(created_at: [start_date..end_date])
etc.) and will likely raise an NotImplementedExrror whenever you call a method that's not supported.

my ideal AbstractAdapter will be a class that lists down all the methods that an adapter requires to
implement to be considered an ActiveRecord adapter, and maybe provide a default implementation that
doesn't do anything more than to raise an exception saying that particular method is not implemented.
that way it behaves more like a proper interface, and any class inheriting from it is a proper
Liskov substitute.

the way it is now, with the number of modules being mixed in, the surface API of AbstractAdapter is
so huge that a person trying to implement a small subset of it for a different backing db is forced
to learn (and trace through) a lot of the implementation, and disable or stub the ones that they don't
need.