implement a session class
i was looking to use faraday, then realized that the gem is probably gonna
be unmaintained by the owner becasue he'll be focusing efforst on a new
gem called hurley.

hurley seems to be almost like faraday, but without the rack dependency.
according to its author, it is what faraday v1.0 would have been without
the baggage.