retrieve the headers
we consume the api payload and pluck out the header names from there.

a naive/poor man's cache is made of the json payload so we don't have to
query the api over and over. we'll need to set a cache expiry for this
one next time.

the easiest way to get the headers is to ask for the document in cell format
and just retrieve the first row. then parse the resulting json. afaict
there is no api to just retrieve the headers directly.