remove caching, add get_list method
it would be better to remove caching here and allow the client app
to implement their own caching. besides, the operations should be
the ones that would need communication with the google servers; if
there is a need to do a search from a returned document for example
(because the client was greedy and tried to get the full doc) then
the client should be searching it in the returned document, and
the gem should not be doing these decisions for them.